+++ src/main/java/gradient/ComputeGradient2d.java
@@ -14,7 +14,7 @@ public class ComputeGradient2d implements ComputeGradient
 	}
 	
 	@Override
-	public void gradientAt( final Localizable location, final float[] derivativeVector )
+	public void gradientAt( final Localizable location, final double[] derivativeVector )
 	{
 		randomAccess.setPosition( location );
 		
@@ -27,7 +27,7 @@ public class ComputeGradient2d implements ComputeGradient
 		randomAccess.bck( 0 );
 		final double p2 = randomAccess.get().getRealDouble();
 		
-		derivativeVector[ 0 ] = (float) ( ( (p1+p3) - (p0+p2) ) / 2.0 );
-		derivativeVector[ 1 ] = (float) ( ( (p2+p3) - (p0+p1) ) / 2.0 );
+		derivativeVector[ 0 ] = ( ( (p1+p3) - (p0+p2) ) / 2.0 );
+		derivativeVector[ 1 ] = ( ( (p2+p3) - (p0+p1) ) / 2.0 );
 	}
 }
